requirements.txt
chromadb-client==0.4.25.dev0
Flask==3.0.2
Flask-SocketIO==5.3.6
huggingface-hub==0.21.4
numpy==1.24.4
openai==0.28.0
pandas==2.0.3
pillow==10.2.0
pymongo==4.6.2
PyPika==0.48.9
python-dotenv==1.0.1
python-socketio==5.11.1
SQLAlchemy==2.0.27
websocket-client==1.7.0
graphlib-backport
eventlet==0.35.2 
gunicorn==21.2.0